Add the WaveformSet.arrange_time_vs_ADC_ranges() method
Also, we are enabling a new feature in WaveformSet.plot_wfs(): now the time range of the heatmap plot can be customized. To do so, plot_wfs() uses the newly added method. Encapsulating this functionality is worth because we are going to also use it in the near future in a new plot() method of the ChannelWSGrid class