Add the ChannelWSGrid.plot() method
This method is very similar to the WaveformSet.plot_wfs() method but does not need to take a grid of waveform-indices as an input since the subplots are spatially ordered according to the self.ChMap attribute. In addition, this plot method includes the calibration mode, where instead of plotting waveforms, the calibration histograms of each channel are plotted.