Decouple plotting-related stuff from data_classes/WaveformSet.py
The plot_calibration_histogram() method of WaveformSet.py is still there. This one is not supported and may not work, and will be deleted soon if we don't need to recycle the code of such method anytime soon. Also, the imports from plotting/plot_utils.py to plotting/plot.py are done function by function. For now, I find it more easy to track what we are importing this way, but this could be replaced by one import line using the wildcard * if needed.