Split the former WfAna class into an abstract base class and BasicWfAna
The new WfAna class should serve as a base class to derive analysis classes. It is now more general since it only contains two attributes: a dictionary which encapsulates the input parameters of an analysis and another dictionary which encapsulates the result of the analysis. The features that were present in the former WfAna class are now part of a BasicWfAna class, which inherits from the new WfAna