Upload the LED-calibration analysis (#45)
* Rename file in src/waffles/np04_utils

* Add the flatten_2D_list() function to src/waffles/utils/wf_maps_utils.py

* Add flat APA maps to data/ProtoDUNE_HD_APA_maps.py

* Move ProtoDUNE_HD_APA_maps.py from data/ to data/np04/

* Add a function to get an iterator value for the flattened APA maps

This solution is inefficient for channel-maps with a very high number of channels. A more efficient solution should be implemented.

* Minor change to src/waffles/input/raw_ROOT_reader.py

* Create a folder for my LED-calibration analysis in np04_analysis/

* Add the map of LED-configurations to channels

This map gives which data (i.e. which LED configuration) should be used to calibrate each channel. This map stores such data for every NP04 APA and every considered overvoltage (for PDEs in [40%, 45%, 50%]).

* Add a folder to store ancillary data for the different calibration batches

* Add a folder for the first calibration batch

* Add map of LED-configuration for each run number in the 1st calib. batch

* Nest the map added in the previous commit in a batch-dependent map

* Add a map of excluded channels for the first calibration batch

Excluded channels shall not be fit in the analysis stage because the data is not well-formed enough

* Nest the excluded-channels map in a batch-dependent map

* Add analysis notebook to src/waffles/np04_analysis/LED_calibration

* Add visualiser notebook to src/waffles/np04_analysis/LED_calibration

* Move the ProtoDUNE_HD_APA_maps.py file inside src/waffles/

* Propagate the changes of the latest commit

This commit and previous one are a quick fix for an ImportError which appeared when trying to import APA_map from data.np04.ProtoDUNE_HD_APA_maps, in analysis.ipynb

* Fix bug in analysis.ipynb

* Fix another bug in analysis.ipynb

* Upload the results dataframe

* Add a minor change to visualiser.ipynb

* Stop ignoring jupyter notebooks