Enable the truncation of every waveform in a WaveformSet (#69)
* PEP8 changes to BasicWfAna.py

* PEP8 change to ChannelMap.py

* PEP8 changes to Map.py

* PEP8 changes to PeakFindingWfAna.py

* PEP8 changes to UniqueChannel.py

* PEP8 changes to Waveform.py

* PEP8 changes to WaveformAdcs.py

* Add the starting_tick attribute to the Waveform class

* Propagate change from commit 1d70ec to the WaveformSet ROOT reader

* Propagate change from commit 1d70ec to the WaveformSet HDF5 reader

In commit 1d70ec3560e7aa9f30615708161ecaf24b1bed0f, the attribute 'starting_tick' was added to the Waveform class

* Substitute the WaveformAdcs.__truncate_adcs() method with __slice_adcs()

The new method allows the selection of the end point of the truncation/slice, which is convenient for the homogeneous-slicing of a WaveformSet which will be introduced in the near future.

* Propagate the change from commit 506e96 to raw_hdf5_reader.py

* Propagate the change from commit 506e96 to raw_root_reader.py

* Fix bug in the plot_WaveformSet() function of plotting/plot.py

This reference to get_map_of_wf_idcs() was not updated when such functionality was moved from a WaveformSet method to a function defined in utils/wf_maps_utils.py

* Add the reset_mean_waveform() method to the WaveformSet class

* Add the truncate_waveforms_in_WaveformSet function to filtering_utils.py

This function is particularly useful for heavy WaveformSet objects (p.e. a waveform set coming from a full-streaming data-taking) for which we are only interested in a narrow time window.

* Minor change to WaveformAdcs.py

* Implement an __slice_adcs() overriding method for the Waveform class

* Propagate change from commit 8e271b to truncate_waveforms_in_WaveformSet

With this commit and commit 8e271bf22f55f66e282a8999e7584fdbc15ed359, the starting_tick attribute of the Waveform objects which are truncated using the truncate_waveforms_in_WaveformSet() function is updated accordingly.