Fix bug in fit_peaks/fit_peaks_utils.py
For an small enough product of percentage_step times len(signal), math.floor() could evaluate to 0 thus adding no points to the running sum, which in turn will never reach the value of len(signal). In this case, the fReachedEnd flag would never be set to True, resulting in an infinite `while`loop.