Two changes to WafflesAnalysis.__steering_file_meets_requirements()
The first one has to do with a new requirement for the steering file. There should be a 'parameters_is_file' parameter alongside the 'name' and 'parameters' parameters. The second one has to do with the decoupling of the functionality of the analysis-class name checking to a new helper method called __check_analysis_class().