Enforce a common workflow for all Waffles analysis (#71)
* Fix bug in fit_peaks/fit_peaks_utils.py

For an small enough product of percentage_step times len(signal), math.floor() could evaluate to 0 thus adding no points to the running sum, which in turn will never reach the value of len(signal). In this case, the fReachedEnd flag would never be set to True, resulting in an infinite `while`loop.

* Remove residual lines

* Add a docstring for the WafflesAnalysis class

* PEP8 changes to WafflesAnalysis.py

* Add a couple of new exceptions derived from WafflesBaseException

* Add the __check_file_or_folder_exists() static method to WafflesAnalysis

* Add __steering_file_meets_requirements static method to WafflesAnalysis

* Add a couple of imports to WafflesAnalysis.py

* Add analysis_folder_meets_requirements static method to WafflesAnalysis

* Delete the `if __name__ == "__main__"` condition

For now, main.py will always be executed as an script. There is no need to separate the execution logic from other potential definitions in main.py.

* Extend documentation of the analysis_folder_meets_requirements() method

* Require the existence of a first analysis stage

This change affects the analysis_folder_meets_requirements() method of the WafflesAnalysis class.

* Minor change to WafflesAnalysis.py

* Enforce the name of the analysis to follow the format 'Analysis<i>'

This change affects the __steering_file_meets_requirements() method of WafflesAnalysis

* Two changes to WafflesAnalysis.__steering_file_meets_requirements()

The first one has to do with a new requirement for the steering file. There should be a 'parameters_is_file' parameter alongside the 'name' and 'parameters' parameters. The second one has to do with the decoupling of the functionality of the analysis-class name checking to a new helper method called __check_analysis_class().

* Add a couple of new exceptions to src/waffles/Exceptions.py

* Add an __init__.py file to src/waffles/core/

* Add an utils.py file to src/waffles/core/

This file hosts utilities which should be used in src/waffles/core/main.py

* Redefine the arguments() abstract method of the WafflesAnalysis class

It is now called define_arguments()

* Redefine the base_arguments() method of the WafflesAnalysis class

It is now called __define_arguments()

* Add the get_ordered_list_of_analyses method to src/waffles/core/utils.py

* Add a verbose parameter to the get_ordered_list_of_analysis() function

This function is defined in src/waffles/core/utils.py

* Require that the params default file is a .yml instead of a .py file

* Minor change to the use_steering_file() function of core/utils.py

* Add a new exception derived from WafflesBaseExeption to Exceptions.py

* Add the empty_string_to_None() function to src/waffles/core/utils.py

This simple tool will be used in the following commits

* Add the __purge_parameters_dictionary function to waffles/core/utils.py

This simple tool will be used in the following commits

* Add the __build_parameters_dictionary_from_shell_string() function

It is a helper function added to src/waffles/core/utils.py

* Add the __build_parameters_dictionary_from_file() function to utils.py

It is a helper function added to src/waffles/core/utils.py

* Add the build_parameters_dictionary() function to core/utils.py

* Rename LED_calibration analysis class from 'analysis' to 'Analysis1'

This change is required to meet the WafflesAnalysis requirements

* Rename the LED calib. analysis file from 'analysis.py' to 'Analysis1.py'

* Rename the LED calib. parameters file from 'params.py' to 'params.yml'

* Create a configs folder within the LED_calibration folder

* Move the calibration_batches/ folder into the configs/ folder

* Append the pydantic package to the requirements file

* Propagate the change introduced in commit 3e4f8c

* Add a default steering file to the LED_calibration analysis

* Add a default params file to the LED_calibration analysis

* Add documentation and style changes to WafflesAnalysis.py

* Add missing import statements to WafflesAnalysis.py

* Define a pydantic BaseModel for the input parameters of waffles analyses

* Minor changes to the __steering_file_meets_requirements() static method

This change only affects WafflesAnalysis.py

* Move the __check_file_or_folder_exists() helper method to core/utils.py

* Propagate the change introduced in commit e82a26

The full hash of the referenced commit is e82a261c538bade2aedfb12885852ff2f708709e

* Remove the leading underscores of __check_file_or_folder_exists function

It was inducing some error related with the name mangling in WafflesAnalysis

* Changes to src/waffles/core/main.py

* Add some import statements to the LED_calibration imports.py file

* Changes to LED_calibration/utils.py

* Changes to the unique analysis class of the LED-calibration analysis

This change only affects LED_calibration/Analysis1.py