Fix bug in __build_parameters_dictionary_from_file() of core/utils.py
If this function received the name of an empty YAML files, it returned None instead of an empty dictionary. This led to a TypeError later on when the build_parameters_dictionary() function queried the length of the returned dictionary. As of now, if the specified YAML file is empty, then an empty dictionary is returned.