Add the feature to overwrite parameters in the steering file (#87)
* Fix bug in __build_parameters_dictionary_from_file() of core/utils.py

If this function received the name of an empty YAML files, it returned None instead of an empty dictionary. This led to a TypeError later on when the build_parameters_dictionary() function queried the length of the returned dictionary. As of now, if the specified YAML file is empty, then an empty dictionary is returned.

* Fix bug in src/waffles/core/utils.py

* Handle YAML ParserError in __build_parameters_dictionary...() function

* Migrate check_analysis_class() func. from WafflesAnalysis to utils.py

The WafflesAnalysis.__check_analysis_class() static method has been migrated to WafflesAnalysis.py. This is needed, together with some more migrations, in order to avoid a circular import between core/utils.py and data_classes/WafflesAnalysis.py

* Changes to the __steering_file_meets_requirements() static method

The required structure for the steering file is now different. This is required in order to let the user overwrite, in the steering file, certain parameters coming from the parameters file.

* Documentation changes to src/waffles/core/utils.py

* Migrate steering_file_meets_...() func. from WafflesAnalysis to utils.py

The WafflesAnalysis.__steering_file_meets_requirements() static method has been migrated to WafflesAnalysis.py. This is needed, together with another migration, in order to avoid a circular import between core/utils.py and data_classes/WafflesAnalysis.py

* Migrate analysis_folder_meet...() func. from WafflesAnalysis to utils.py

The WafflesAnalysis.__analysis_folder_meets_requirements() static method has been migrated to WafflesAnalysis.py. This is needed in order to avoid a circular import between core/utils.py and data_classes/WafflesAnalysis.py

* Propagate change from commit d596c259ecc59bc27ffc9ddb36ad2ac39f6f9026

* Fix circular import error

* Minor change to the steering_file_meets_requirements func. documentation

* Changes to the get_ordered_list_of_analyses() function of core/utils.py

The structure of list delivered by this function is now different. This is required in order to let the user overwrite, in the steering file, certain parameters coming from the parameters file.

* Propagate changes from commits fd14eb and 8b5fb9

This is required in order to let the user overwrite, in the steering file, certain parameters coming from the parameters file.

* Warn the user when additional parameters given to main.py are ignored

* Propagate changes from commit fd14eb to the led_calibration analysis

* Propagate changes from commit fd14eb to tau_slow_convolution analysis

* Remove unused import from src/waffles/core/main.py

* Remove unused imports from src/waffles/data_classes/WafflesAnalysis.py