Fix subsampling bug in the WaveformSet_from_hdf5_file() function
The total number of saved waveforms was scaled down by the selected subsampling rate. P.e. setting wvfm_count to 100 and subsample to 2 would return a WaveformSet with 50 waveforms, instead of the first 100 every other waveforms. According to the documentation of the wvfm_count parameter, it should give the total number of waveforms to save.