Fix subsampling bug and reduce duplicated code in HDF5 reader (#93)
* Fix subsampling bug in the WaveformSet_from_hdf5_file() function

The total number of saved waveforms was scaled down by the selected subsampling rate. P.e. setting wvfm_count to 100 and subsample to 2 would return a WaveformSet with 50 waveforms, instead of the first 100 every other waveforms. According to the documentation of the wvfm_count parameter, it should give the total number of waveforms to save.

* Encapsulate the truncation-to-minimum code in a function

It will be needed in the next commits to remove duplicated code in `input_output/raw_hdf5_reader.py`

* Use the function introduced in commit 592b11 in `raw_root_reader.py`

* Reduce duplicated code in WaveformSet_from_hdf5_file()