Add the StoreWfAna class
This class inherits from WfAna. It implements a dummy analysis which simply passes the given input parameters as the result of the StoreWfAna 'analysis'. It may be helpful to add the trigger primitives coming from the DAQ as if they were the result of an analysis performed in waffles. In this way, the TPs and the results of further offline analyses are available at the same level in waffles