Specify that the adcs attribute of a WaveformAdcs may contain floats
Their entries are not required to be integers anywhere in /waffles/src/waffles/data_classes/. A thorough analysis of the rest of subfolders in /waffles/src/waffles/ was not done. Relaxing this 'restriction' from int to int/float is necessary to introduce methods which can potentially modify the adcs attribute of WaveformAdcs objects, p.e. subtracting a float baseline, resulting in a float adcs array.