LED-calibration development (#125)
* Minor change to get_input_params_model() of led_calibration/Analysis1.py

* Add the WfAna-derived class WindowBaseliner

This class is meant to be an analyser class whose only purpose is to compute the baseline of a waveform given a fixed time window relative to that waveform.

* Add the WfAna-derived class WindowIntegrator

This class is meant to be an analyser class whose only purpose is to integrate a time-fixed window of the given waveform.

* Minor change to WindowBaseliner.py

* Add LED-calibration configuration files

* WindowBaseliner analyzer now returns the std of the baseline

* Add the filter for the LED-calibration selection

* Add the get_average_baseline_std() function to utils.py

* Promote boolean-annotation check for the return type to shallower level

* Propagate changes from commit 5e02f2 to WaveformSet.py

At the moment of writing, there are no occurrences of "check_well_formedness_of_generic_waveform_function" in the main branch of waffles/src other than those in WaveformSet.py. I.e. the propagation of the changes of the mentioned commit should only affect WaveformSet.py

* Fix a couple of typos

* Specify that the adcs attribute of a WaveformAdcs may contain floats

Their entries are not required to be integers anywhere in /waffles/src/waffles/data_classes/. A thorough analysis of the rest of subfolders in /waffles/src/waffles/ was not done. Relaxing this 'restriction' from int to int/float is necessary to introduce methods which can potentially modify the adcs attribute of WaveformAdcs objects, p.e. subtracting a float baseline, resulting in a float adcs array.

* Use WaveformSet.reset_mean_waveform() where appropriate

* Add a setter method for the adcs attribute of the WaveformAdcs class

* Add the WaveformSet.apply() method

* Add the subtract_baseline() function

* Fix bug introduced in commit 9218ae

* Add missing entry to the BasicWfAna input parameters

* Separate the calibration-histogram-computation from the __init__ method

* Add the compute_calib_histos() method of the ChannelWsGrid class

This method is convenient for cases when one cannot compute the channel-wise calibration histograms at the initialization stage of the ChannelWsGrid object, but needs to do so later.

* Add some type hints to fit_peaks.py

* Make the half_points_to_fit argument optional in a couple of functions

* Propagate change from commit 27fadd

This change is not strictly necessary though

* Delete unused import

* Encapsulate the independent-gaussians fit into a separate function

This is convenient since in further commits we want to introduce another way to do the gaussian fit (a correlated fit with the gain as a fitting parameter) in parallel with the independent-gaussians type of fit

* Add a parameter to choose how to fit the calibration histogram peaks

* Add the correlated_sum_of_gaussians() function to numerical_utils.py

* Add the get_pulse_window_limits() function

This function will be used to infer, channel by channel, the integration limits for the LED-calibration analysis

* Add the __fit_correlated_gaussians_to_calibration_histogram() function

This function gives an alternative way (to the one which uses one independent gaussian per peak) to fit the charge calibration histogram

* Style changes to the __spot_first_peaks_in_CalibrationHistogram function

* Add the return_last_addition_if_fail parameter to the __spot_...() func.

* Add a parameter to the fit_peaks_of_CalibrationHistogram() function

The added parameter is return_last_addition_if_fail

* Add a parameter to the fit_peaks_of_ChannelWsGrid() function

The added parameter is return_last_addition_if_fail

* Style change to the fit_peaks_of_CalibrationHistogram() function

* Implement the fit_type=='correlated_gaussians' case

This implementation affects the fit_peaks_of_CalibrationHistogram() function

* Add a couple of parameters to the fit_peaks_of_ChannelWsGrid() function

The added parameters are std_increment_seed_fallback and ch_span_fraction_around_peaks

* Add the plot_sum_of_gaussians parameter to plot_CalibrationHistogram()

This is needed to visualize the guassian fit of the calibration histogram when the correlated-type of fit was used

* Add the plot_sum_of_gaussians parameter to plot_ChannelWsGrid()

* Update the LED-calibration Analysis1 class

The used import statements are updated accordingly

* Update the utilities associated to the LED-calibration analysis

* Add entries for runs of batch 4 acquired at nominal V-gain

* Propose aimed_channels for the newly added batch-4 runs

This proposal is based on previous batches

* Propose no-excluded channels for the newly added batch-4 runs

* Add entries for runs of batch 5 acquired at nominal V-gain

* Propose aimed_channels for the newly added batch-5 runs

* Propose no excluded channels for the newly added batch-5 runs

* Add entries for runs of batch 6 acquired at nominal V-gain

* Propose aimed_channels for newly added batch-6 runs

* Propose no excluded channels for the newly added batch-6 runs

* Add entries for runs of batch 7 acquired for APAs 2, 3 and 4 at 3 PDEs

* Propose aimed_channels for newly added batch-7 runs

* Propose excluded channels for newly added batch-7 runs

* Exclude disconnected channels from APA 2 in batches 4 and 5

* Change name of format_channel_number() function

* Add the split_channel_number() function

* Add the arrange_dictionary_of_endpoints_and_channels() function

* Add the dump_object_to_pickle() function

* Update batch_pickle_generator.ipynb

* Add the next_subsample() function

* Fix bug in batch_pickle_generator.ipynb

* Minor changes to batch_pickle_generator.ipynb

* Use the next_subsample() function in batch_pickle_generator.ipynb