Excluding channel 105-12 of APA 1 for batches 4 and 5 at PDE 45%
The waveforms for this channel (run 28633 and 28979) have lots of points (approximately around 10% of the them) which fall to ~4100 ADCs or to ~0 ADCs. They are randomly distributed along the 1024 points of the waveform. Probably, these points can be easily discarded and substituted with an neighbouring average. This channel is kept aside for the moment, though.