Update the excluded channels for APA 1 data of batch 1
After testing the coarse_selection_for_led_calibration() cut there is typically not even one waveform left for channel 105-12, and so, it has been excluded for all three PDEs of batch 1 and APA 1. With the other changes that have been introduced lately, we can recover channels 107-00, 107-02 and 107-05, but we should exclude channel 105-00.