LED-calibration development (#138)
* Minor change to get_input_params_model() of led_calibration/Analysis1.py

* Add the WfAna-derived class WindowBaseliner

This class is meant to be an analyser class whose only purpose is to compute the baseline of a waveform given a fixed time window relative to that waveform.

* Minor change to WindowBaseliner.py

* WindowBaseliner analyzer now returns the std of the baseline

* Use WaveformSet.reset_mean_waveform() where appropriate

* Add a setter method for the adcs attribute of the WaveformAdcs class

* Add the WaveformSet.apply() method

* Fix bug introduced in commit 9218ae

* Propagate change from commit 27fadd

This change is not strictly necessary though

* Add the correlated_sum_of_gaussians() function to numerical_utils.py

* Add the plot_sum_of_gaussians parameter to plot_CalibrationHistogram()

This is needed to visualize the guassian fit of the calibration histogram when the correlated-type of fit was used

* Update the LED-calibration Analysis1 class

The used import statements are updated accordingly

* Add temporal_copy_directory parameter to batch_pickle_generator.ipynb

* Add code cell to check the completeness of the downloaded pickles

* Separate pickled runs into different folders batch- and pde-wise

* Add code cell to separated pickle'd WaveformSets into organized folders

* Skip configurations for which there are no runs acquired

* Skip the runs for which there are no aimed channels

* Add some checks to the parse_numeric_list()

* Fix bug in the parse_numeric_list() function

* Substitute get_input_folderpath() with get_input_filepaths_for_run()

* Minor changes to pickle_file_reader.py

* Add note to Analysis1.py

* Use the new get_input_filepaths_for_run() function in Analysis1.py

* Delete the read_data() function and use directly WaveformSet_..._file()

* Add the truncate_wfs_to_minimum parameter for batch 1 (APAs 1&2)

* Excluding channel 105-12 of APA 1 for batches 4 and 5 at PDE 45%

The waveforms for this channel (run 28633 and 28979) have lots of points (approximately around 10% of the them) which fall to ~4100 ADCs or to ~0 ADCs. They are randomly distributed along the 1024 points of the waveform. Probably, these points can be easily discarded and substituted with an neighbouring average. This channel is kept aside for the moment, though.

* Exclude channel 111-31 of APA 3 for batches 4-7

This channel, which was already excluded for batches 1-3, won't show two resolved peaks in the charge histogram when using the same analysis as for the rest of the APA 3 channels. It needs deeper investigation. For the moment it is kept aside.

* Handle possible ill-formedness error in selection_for_led_calibration()

* Rename the selection_for_led_calibration() function

* Fix inaccuracies in filtering_utils.py

* Add the coarse_selection_for_led_calibration() function

The addition of this filter is motivated by some APA 1 data (specially that of channel 105-12) for which some ADC points randomly explode to values such as ~4100 ADCs or ~0 ADCs.

* Update the excluded channels for APA 1 data of batch 1

After testing the coarse_selection_for_led_calibration() cut there is typically not even one waveform left for channel 105-12, and so, it has been excluded for all three PDEs of batch 1 and APA 1. With the other changes that have been introduced lately, we can recover channels 107-00, 107-02 and 107-05, but we should exclude channel 105-00.

* Update the excluded channels for APA 2 data of batch 1

Channel 109-20 at 40% PDE is recovered, but channel 109-33 is excluded. At 50%, the channel 109-15 is excluded.

* Update the excluded channels for APA 3 data of batch 1

Channel 111-15 is recovered for PDE 40%

* Update the excluded channels for APA 4 data of batch 1

Channel 113-02 is recovered for PDE 40%

* Add the EmptyCalibrationHistogram exception

* Style changes to CalibrationHistogram.py

* Raise an exception if an empty CalibrationHistogram is initialized

* Handle the EmptyCalibrationHistogram exception in compute_calib_histos()

* Skip channels with no calib. histo. in fit_peaks_of_ChannelWsGrid()

* Style change to src/waffles/plotting/plot.py

* Avoid raising an exception if a calibration histogram is not available

* Update the excluded channels for APA 1 data of batch 2 at 40% PDE

Most of them are due to the fact that there are no waveforms for those channels. Is it possible that they were unbiased during the data-taking?

* Update the excluded channels for APA 1 data of batch 2 at 45% PDE

* Update the excluded channels for APA 1 data of batch 2 at 50% PDE

Most of them are due to the fact that there are no waveforms for those channels. Is it possible that they were unbiased during the data-taking?

* Update the excluded channels for APAs 3 & 4 data of batch 2

* Update the excluded channels for batch 3

APA 1 was not acquired during this measurements batch. APAs 2 and 3 are not affected. Some channels for APA 4 are recovered at PDEs 45% and 50%.

* Update the excluded channels for APA 1 data of batch 4

All of the endpoint 107 channels are excluded. These channels need further investigation, since their charge histogram hardly display separated peaks.

* Update the excluded channels for APAs 2, 3 and 4 of batch 4

APAs 3 and 4 are not affected.

* Add the excluded channels for batch 5

* Add the excluded channels for batch 6

* Add the excluded channels for APA 2 data of batch 7

A finer tuning of the input parameters for this case may let us recover some channels

* Add the excluded channels for APA 3 data of batch 7

These should be fine tuned yet

* Add the excluded channels for APA 4 data of batch 7

These should be fine tuned yet. I.e. we need to investigate whether there are alternative runs which give better charge histograms for the currently excluded channels.

* Use the coarse_selection_for_led_calibration() function in Analysis1.py

* Parameters 'baseline_i_up' & 'signal_i_up' become APA-wise dictionaries

Also, a local variable 'aux_time_increment', which is an APA-wise dictionary, is introduced in Analysis1.write_output() just for plotting purposes

* Use the verbosity parameter where needed in led_calibration/Analysis1.py

* Rename some local variables of Analysis1.analyze()

* Update led_calibration/params.yml including the recently added params.

* Add a CSV file including the dates for each batch

* Update led_calibration/scripts/visualiser.ipynb

* Rename function get_average_baseline_std()

* Remove duplicated chunk of code

* Resolve conflicts in plotting/plot.py