Changes to the `save_data_to_dataframe()` function
This changes include changing the name of the former `data` parameter, which is now named `packed_gain_and_snr`, adding the `packed_integration_limits` parameter and simplifying the check of the existing columns in the running dataframe. If provided, the integration limits for each channel are now saved to the output dataframe.