Change to the `add_SPE_info_to_output_dictionary()` function
If the 1-PE peak is not found (because the calibration did not provide it) or the limits calculation fail, then the entry for the SPE related entries in the output-data dictionary is not set to 'unavailable' anymore, but it is simply not added. This is more convenient because further functions that are applied further down the analysis pipeline (and that will be included in the following commits) handle missing entries, but do not handle an 'unavailable' entry. For the moment it is more convenient to not separate different handling cases.